# Added bulkInsert() support to API and ContentProvider for 2000% faster Note inserts