@marcardar 2016-03-02
I'm going to split this PR into smaller PRs